# api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "schedule-controller",
      "description": "Schedule Controller"
    },
    {
      "name": "tennis-court-controller",
      "description": "Tennis Court Controller"
    }
  ],
  "paths": {
    "/tenniscourts-api/schedule/add": {
      "post": {
        "tags": [
          "schedule-controller"
        ],
        "summary": "addScheduleTennisCourt",
        "operationId": "addScheduleTennisCourtUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "createScheduleRequestDTO",
            "description": "createScheduleRequestDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateScheduleRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/tenniscourts-api/schedule/findByScheduleId": {
      "get": {
        "tags": [
          "schedule-controller"
        ],
        "summary": "findByScheduleId",
        "operationId": "findByScheduleIdUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "query",
            "description": "scheduleId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ScheduleDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/tenniscourts-api/schedule/findSchedulesByDates": {
      "get": {
        "tags": [
          "schedule-controller"
        ],
        "summary": "findSchedulesByDates",
        "operationId": "findSchedulesByDatesUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "endDate",
            "in": "query",
            "description": "endDate",
            "required": true,
            "type": "string",
            "format": "date"
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "startDate",
            "required": true,
            "type": "string",
            "format": "date"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ScheduleDTO"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/tenniscourts-api/tennis-court/add": {
      "post": {
        "tags": [
          "tennis-court-controller"
        ],
        "summary": "addTennisCourt",
        "operationId": "addTennisCourtUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "tennisCourtDTO",
            "description": "tennisCourtDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TennisCourtDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/tenniscourts-api/tennis-court/findTennisCourtById": {
      "get": {
        "tags": [
          "tennis-court-controller"
        ],
        "summary": "findTennisCourtById",
        "operationId": "findTennisCourtByIdUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tennisCourtId",
            "in": "query",
            "description": "tennisCourtId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TennisCourtDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/tenniscourts-api/tennis-court/findTennisCourtWithSchedulesById": {
      "get": {
        "tags": [
          "tennis-court-controller"
        ],
        "summary": "findTennisCourtWithSchedulesById",
        "operationId": "findTennisCourtWithSchedulesByIdUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tennisCourtId",
            "in": "query",
            "description": "tennisCourtId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TennisCourtDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
    "CreateScheduleRequestDTO": {
      "type": "object",
      "properties": {
        "startDateTime": {
          "type": "string",
          "format": "date-time"
        },
        "tennisCourtId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "CreateScheduleRequestDTO"
    },
    "ScheduleDTO": {
      "type": "object",
      "properties": {
        "endDateTime": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "startDateTime": {
          "type": "string",
          "format": "date-time"
        },
        "tennisCourt": {
          "$ref": "#/definitions/TennisCourtDTO"
        },
        "tennisCourtId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "ScheduleDTO"
    },
    "TennisCourtDTO": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "tennisCourtSchedules": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ScheduleDTO"
          }
        }
      },
      "title": "TennisCourtDTO"
    }
  }
}